+++ rust/swift_bindings/ChromaDemos/EphemeralChroma/EphemeralChroma/EphemeralChroma.entitlements
