+++ rust/swift_bindings/ChromaDemos/EphemeralChroma/EphemeralChromaTests/EphemeralChromaTests.swift
@@ -1,0 +1,16 @@
+//
+//  EphemeralChromaTests.swift
+//  EphemeralChromaTests
+//
+//  Created by Nicholas Arner on 5/21/25.
+//
+
+import Testing
+
+struct EphemeralChromaTests {
+
+    @Test func example() async throws {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+    }
+
+}
