+++ rust/swift_bindings/ChromaDemos/PersistentChroma/PersistentChroma/PersistentChromaApp.swift
@@ -1,0 +1,17 @@
+//
+//  PersistentChromaApp.swift
+//  PersistentChroma
+//
+//  Created by Nicholas Arner on 5/21/25.
+//
+
+import SwiftUI
+
+@main
+struct PersistentChromaApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
