+++ rust/swift_bindings/ChromaDemos/PersistentChroma/PersistentChromaUITests/PersistentChromaUITests.swift
@@ -1,13 +1,13 @@
 //
-//  ChromaDemoUITests.swift
-//  ChromaDemoUITests
+//  PersistentChromaUITests.swift
+//  PersistentChromaUITests
 //
-//  Created by Nicholas Arner on 5/14/25.
+//  Created by Nicholas Arner on 5/21/25.
 //
 
 import XCTest
 
-final class ChromaDemoUITests: XCTestCase {
+final class PersistentChromaUITests: XCTestCase {
 
     override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
