+++ rust/swift_bindings/ChromaDemos/PersistentChroma/PersistentChromaUITests/PersistentChromaUITestsLaunchTests.swift
@@ -1,0 +1,33 @@
+//
+//  PersistentChromaUITestsLaunchTests.swift
+//  PersistentChromaUITests
+//
+//  Created by Nicholas Arner on 5/21/25.
+//
+
+import XCTest
+
+final class PersistentChromaUITestsLaunchTests: XCTestCase {
+
+    override class var runsForEachTargetApplicationUIConfiguration: Bool {
+        true
+    }
+
+    override func setUpWithError() throws {
+        continueAfterFailure = false
+    }
+
+    @MainActor
+    func testLaunch() throws {
+        let app = XCUIApplication()
+        app.launch()
+
+        // Insert steps here to perform after app launch but before taking a screenshot,
+        // such as logging into a test account or navigating somewhere in the app
+
+        let attachment = XCTAttachment(screenshot: app.screenshot())
+        attachment.name = "Launch Screen"
+        attachment.lifetime = .keepAlways
+        add(attachment)
+    }
+}
